Use different parameter files for Q-Learning and Policy Gradient
fixes #318
