+++ ros_ws/src/autonomous/reinforcement_learning/scripts/reinforcement_learning_node.py
@@ -7,24 +7,26 @@ from drive_msgs.msg import drive_param
 
 import torch
 
-from parameters import *
+from topics import TOPIC_DRIVE_PARAMETERS, TOPIC_SCAN
 
 '''
 Abstract class for methods that are used both
 during training and during driving.
 '''
 class ReinforcementLearningNode():
-    def __init__(self):
+    def __init__(self, actions, laser_sample_count):
         self.scan_indices = None
+        self.laser_sample_count = laser_sample_count
+        self.actions = actions
         self.drive_parameters_publisher = rospy.Publisher(
             TOPIC_DRIVE_PARAMETERS, drive_param, queue_size=1)
         rospy.Subscriber(TOPIC_SCAN, LaserScan, self.on_receive_laser_scan)
 
     def perform_action(self, action_index):
-        if action_index < 0 or action_index >= len(ACTIONS):
+        if action_index < 0 or action_index >= len(self.actions):
             raise Exception("Invalid action: " + str(action_index))
 
-        angle, velocity = ACTIONS[action_index]
+        angle, velocity = self.actions[action_index]
         message = drive_param()
         message.angle = angle
         message.velocity = velocity
@@ -32,7 +34,7 @@ class ReinforcementLearningNode():
 
     def convert_laser_message_to_tensor(self, message):
         if self.scan_indices is None:
-            self.scan_indices = [int(i * (len(message.ranges) - 1) / (LASER_SAMPLE_COUNT - 1)) for i in range(LASER_SAMPLE_COUNT)]  # nopep8
+            self.scan_indices = [int(i * (len(message.ranges) - 1) / (self.laser_sample_count - 1)) for i in range(self.laser_sample_count)]  # nopep8
 
         values = [message.ranges[i] for i in self.scan_indices]
         values = [v if not math.isinf(v) else 100 for v in values]
