+++ ros_ws/src/autonomous/reinforcement_learning/scripts/topics.py
@@ -1,0 +1,5 @@
+TOPIC_DRIVE_PARAMETERS = "/input/drive_param/autonomous"
+TOPIC_SCAN = "/scan"
+TOPIC_CRASH = "/crash"
+TOPIC_GAZEBO_MODEL_STATE = "/gazebo/model_states"
+TOPIC_EPISODE_RESULT = "/qlearning/episode_result"
